Add & Fix 104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -12,7 +12,7 @@
  */
 void fibonacci(int num)
 {
-	long long int num1, num2, placeholder;
+	long unsigned int num1, num2, placeholder;
 
 	num1 = 1;
 	num2 = 2;
